long to int conversion

+++ src/main/java/com/scalx/devnews/repository/ArticleRepository.java
@@ -5,11 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
 
     public Article findByTitle(String title);
 
-    public Article findById(long id);
+    public Article findById(int id);
 
     public Article findByArticleLink(String link);
 }
